+++ fennec/settings/common.py
@@ -192,8 +192,8 @@ THIRD_PARTY_APPS = (
 LOCAL_APPS = (
     'fennec.apps.dashboard',
     'fennec.apps.authentication',
-    'fennec.apps.diagram',
-    'fennec.apps.versioncontroll',
+    'fennec.apps.metamodel',
+    'fennec.apps.repository',
 )
 
 # See: https://docs.djangoproject.com/en/dev/ref/settings/#installed-apps
